src/Utilities/Hashing/SHA256Hasher.cs
using System.Security.Cryptography;
using System.Text;

namespace Utilities.Hashing
{
    public static class SHA256Hasher
    {
        public static string GetHash(string input)
        {
            StringBuilder stringBuilder = new StringBuilder();

            using (SHA256 hash = SHA256.Create())
            {
                Encoding encoding = Encoding.UTF8;

                byte[] result = hash.ComputeHash(encoding.GetBytes(input));

                foreach (byte b in result)
                    stringBuilder.Append(b.ToString("x2"));
            }

            return stringBuilder.ToString();
        }
    }
}
